DEP-6140 Updated library version to use play 2.9, major change to library functionality to accomodate NONCE changes.
