+++ play-29/src/main/scala/uk/gov/hmrc/webchat/utils/ParameterEncoder.scala
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2024 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package uk.gov.hmrc.webchat.utils
+
+import java.net.{URLDecoder, URLEncoder}
+
+import play.api.libs.json.Json
+
+object ParameterEncoder {
+  def encodeStringList(list: Seq[String]): String = {
+    URLEncoder.encode(Json.toJson(list).toString, "UTF-8")
+  }
+  def decodeStringList(encodedList: String): Seq[String] = {
+    val decoded = URLDecoder.decode(encodedList, "UTF-8")
+    Json.parse(decoded).as[Seq[String]]
+  }
+}
