+++ play-29/src/main/twirl/uk/gov/hmrc/webchat/views/HMRCEmbeddedView.scala.html
@@ -1,0 +1,27 @@
+@*
+ * Copyright 2024 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *@
+
+@import views.html.helper.CSPNonce
+@import uk.gov.hmrc.webchat.config.WebChatConfig
+@import play.api.mvc.RequestHeader
+
+@this(appConfig: WebChatConfig)
+
+@()(implicit request: RequestHeader)
+
+<script @{CSPNonce.attr} language="javascript" id="hmrc-webchat-tag" src=@{appConfig.hmrcSkinJSUrl}></script>
+<link @{CSPNonce.attr} rel="stylesheet" href='@{appConfig.hmrcSkinEmbeddedCSSUrl}' type="text/css">
+<div class="dav3"></div>
