+++ play-29/src/main/twirl/uk/gov/hmrc/webchat/views/NuanceView.scala.html
@@ -1,0 +1,32 @@
+@*
+ * Copyright 2024 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *@
+
+@import uk.gov.hmrc.webchat.config.WebChatConfig
+@import uk.gov.hmrc.webchat.models.EncryptedNuanceData
+
+@this(appConfig: WebChatConfig)
+
+@(nuanceData: EncryptedNuanceData)
+
+<script>
+    var nuanceData = {
+        "mdtpdfSessionID" : "@{nuanceData.nuanceSessionId}",
+        "mdtpSessionID" : "@{nuanceData.mdtpSessionId}",
+        "deviceID" : "@{nuanceData.deviceId}"
+    }
+</script>
+
+<script id="webchat-tag" src=@{appConfig.nuanceUrl}></script>
