--- src/main/scala/uk/gov/hmrc/webchat/client/WebChatClient.scala
@@ -1,29 +1,0 @@
-/*
- * Copyright 2023 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package uk.gov.hmrc.webchat.client
-
-import com.google.inject.ImplementedBy
-import play.api.mvc.Request
-import play.twirl.api.Html
-import uk.gov.hmrc.webchat.internal.WebChatClientImpl
-
-@ImplementedBy(classOf[WebChatClientImpl])
-trait WebChatClient {
-  def loadRequiredElements()(implicit request: Request[_]): Option[Html]
-  def loadHMRCChatSkinElement(partialType: String)(implicit request: Request[_]): Option[Html]
-  def loadWebChatContainer(id: String = "HMRC_Fixed_1")(implicit request: Request[_]) : Option[Html]
-}
